+++ src/app/contextual-action-panel/contextual-action-panel.demo.ts
@@ -30,7 +30,7 @@ export class ContextualActionPanelDemoComponent {
    */
   public panelRef?: SohoContextualActionPanelRef<any> | null;
   public closeResult?: string;
-  public title = 'Contextual Action Panel';
+  public title = 'Title using modalSettings title';
 
   /**
    * Constructor.
@@ -77,6 +77,9 @@ export class ContextualActionPanelDemoComponent {
       })
       .afterOpen(() => {
         console.log('After Open Fires');
+      })
+      .beforeClose(() => {
+        console.log('Before close fires');
       })
       .closed(() => {
         console.log('Closed Fires');